Change index router to only show users their own questions and comment out the show router
Users were able to see all questions ever made, so changes were made so
that users could only view their own questions. It was also determined
that the show router is not necessary at the moment but just in case it
will be commented out until it is deemed fully necessary to delete the
code.